+++ routes/admin/products.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const router = express.Router();
+const productsRepo = require("../../repositories/products");
 
 router.get("/admin/products", (req, res) => {});
